% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{MEASURABILITY_LEVELS}
\alias{MEASURABILITY_LEVELS}
\title{Measurability-class levels}
\format{
An object of class \code{character} of length 4.
}
\usage{
MEASURABILITY_LEVELS
}
\description{
Drives the gap-fill ladder (SCOPING §6).
}
\seealso{
Other enums: 
\code{\link{METHOD_LEVELS}},
\code{\link{QC_FLAG_LEVELS}},
\code{\link{STAT_CLASS_LEVELS}},
\code{\link{TIER_LEVELS}},
\code{\link{tier_rank}()},
\code{\link{validate_measurability_class}()},
\code{\link{validate_method}()},
\code{\link{validate_qc_flag}()},
\code{\link{validate_statistical_class}()},
\code{\link{validate_tier}()}
}
\concept{enums}
\keyword{datasets}
