% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{met_variable}
\alias{met_variable}
\title{Look up a single dictionary variable}
\usage{
met_variable(variable)
}
\arguments{
\item{variable}{A single variable name.}
}
\value{
A one-row tibble (see \code{\link[=met_variables]{met_variables()}} for columns).
}
\description{
Look up a single dictionary variable
}
\examples{
met_variable("temperature_2m")
}
\seealso{
Other dictionary: 
\code{\link{met_register_variable}()},
\code{\link{met_variables}()}
}
\concept{dictionary}
