% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{METHOD_LEVELS}
\alias{METHOD_LEVELS}
\title{Production-method levels}
\format{
An object of class \code{character} of length 7.
}
\usage{
METHOD_LEVELS
}
\description{
How a canonical value was produced.
}
\seealso{
Other enums: 
\code{\link{MEASURABILITY_LEVELS}},
\code{\link{QC_FLAG_LEVELS}},
\code{\link{STAT_CLASS_LEVELS}},
\code{\link{TIER_LEVELS}},
\code{\link{tier_rank}()},
\code{\link{validate_measurability_class}()},
\code{\link{validate_method}()},
\code{\link{validate_qc_flag}()},
\code{\link{validate_statistical_class}()},
\code{\link{validate_tier}()}
}
\concept{enums}
\keyword{datasets}
