% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-list.R
\name{site_ids}
\alias{site_ids}
\title{The site IDs in a \code{met_sites} collection}
\usage{
site_ids(sites)
}
\arguments{
\item{sites}{A \code{\link[=met_sites]{met_sites()}} object.}
}
\value{
A character vector of \code{site_id} values, in order.
}
\description{
The site IDs in a \code{met_sites} collection
}
\examples{
site_ids(met_sites(list()))
}
\seealso{
Other site: 
\code{\link{as_met_sites}()},
\code{\link{met_instrument}()},
\code{\link{met_site}()},
\code{\link{met_sites}()},
\code{\link{read_sites_yaml}()},
\code{\link{site_id}()},
\code{\link{site_resolved}()},
\code{\link{write_sites_yaml}()}
}
\concept{site}
