% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{TIER_LEVELS}
\alias{TIER_LEVELS}
\title{Correction-tier levels}
\format{
An object of class \code{character} of length 5.
}
\usage{
TIER_LEVELS
}
\description{
Ordered from least to most corrected. Use \code{\link[=tier_rank]{tier_rank()}} to compare tiers.
}
\seealso{
Other enums: 
\code{\link{MEASURABILITY_LEVELS}},
\code{\link{METHOD_LEVELS}},
\code{\link{QC_FLAG_LEVELS}},
\code{\link{STAT_CLASS_LEVELS}},
\code{\link{tier_rank}()},
\code{\link{validate_measurability_class}()},
\code{\link{validate_method}()},
\code{\link{validate_qc_flag}()},
\code{\link{validate_statistical_class}()},
\code{\link{validate_tier}()}
}
\concept{enums}
\keyword{datasets}
