% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{STAT_CLASS_LEVELS}
\alias{STAT_CLASS_LEVELS}
\title{Statistical-class levels}
\format{
An object of class \code{character} of length 5.
}
\usage{
STAT_CLASS_LEVELS
}
\description{
Drives QC and correction dispatch (SCOPING §3, §6).
}
\seealso{
Other enums: 
\code{\link{MEASURABILITY_LEVELS}},
\code{\link{METHOD_LEVELS}},
\code{\link{QC_FLAG_LEVELS}},
\code{\link{TIER_LEVELS}},
\code{\link{tier_rank}()},
\code{\link{validate_measurability_class}()},
\code{\link{validate_method}()},
\code{\link{validate_qc_flag}()},
\code{\link{validate_statistical_class}()},
\code{\link{validate_tier}()}
}
\concept{enums}
\keyword{datasets}
